Move _Randomize into pipeline_utils and setup v2
